requirements.txt
psycopg[binary,pool]
coinbase==2.1.0
pylint
ipython
pep8
uuid
pytz
pyyaml
python-dotenv
beautifulsoup4
requests
